+++ src/components/Query/Rest.vue
@@ -1,6 +1,16 @@
 <template>
   <div>
-    <rest-query-history class="mb-8" @setRequest="setRequest" table-name="rest"/>
+    <div class="mb-12">
+      <v-btn class="pl-1 mr-2" @click="historyCollapsed = !historyCollapsed" title="Show history">
+        <v-icon>{{ historyCollapsed ? 'mdi-chevron-up' : 'mdi-chevron-down' }}</v-icon>
+        History
+      </v-btn>
+
+      <rest-query-examples @setRequest="setRequest"/>
+      <v-expand-transition>
+        <rest-query-history @setRequest="setRequest" v-if="historyCollapsed"/>
+      </v-expand-transition>
+    </div>
 
     <v-form @submit.prevent="loadData">
       <v-row>
@@ -24,10 +34,7 @@
 
       <v-row>
         <v-col :md="vertical ? 12 : 6" cols="12">
-          <h4 class="pb-1">{{ $t('rest.request-body') }} <span v-if="!canSendBody">({{ $t('rest.disabled') }})</span></h4>
-          <resizable-container v-if="canSendBody" ref="query_body"
-                               :initial-height="vertical ? 200 : 500"
-                               class="mb-4">
+          <resizable-container v-if="canSendBody" :initial-height="vertical ? 200 : 500" class="mb-4">
             <code-editor v-model="requestBody" :external-commands="editorCommands"/>
           </resizable-container>
 
@@ -46,29 +53,27 @@
             </div>
           </div>
 
-          <v-row>
-            <v-col :md="vertical ? 6 : 12" cols="12">
-              <v-btn id="execute_query" :disabled="!formValid" :loading="loading" class="mx-0" color="primary-button"
-                     type="submit">
-                {{ $t('rest.execute-query') }}
-              </v-btn>
-              <button id="reset-form" type="button" class="btn-link ml-2" @click="resetForm">{{ $t('rest.reset-form') }}</button>
-            </v-col>
-            <v-col :class="vertical ? 'text-right' : ''" :md="vertical ? 6 : 12" cols="12">
-              <button id="example-1" type="button" class="btn-link" @click="loadCatExample">{{ $t('rest.example-1') }}
-              </button>
-              <button id="example-2" type="button" class="btn-link ml-2" @click="loadCreateExample">
-                {{ $t('rest.example-2') }}
-              </button>
-              <button id="example-3" type="button" class="btn-link ml-2" @click="loadDeleteExample">
-                {{ $t('rest.example-3') }}
-              </button>
-            </v-col>
-          </v-row>
+          <v-btn id="execute_query" :disabled="!formValid" :loading="loading" class="mx-0" color="primary-button"
+                 type="submit">
+            {{ $t('rest.send-request') }}
+          </v-btn>
+          <v-chip class="ml-2" :class="responseStatusClass" v-if="responseStatus">{{ responseStatus }}</v-chip>
+
+          <br>
+          <v-btn id="reset-form" class="mt-2" @click="resetForm" small text>
+            {{ $t('rest.reset') }}
+          </v-btn>
         </v-col>
 
         <v-col :md="vertical ? 12 : 6" cols="12">
-          <print-pretty :caption="responseCaption" :document="responseBody" :focus="false" class="response mb-4"/>
+          <print-pretty :document="responseBody" :focus="false" class="mb-2"/>
+          <v-btn :disabled="!responseBody || responseBody.length === 0"
+                 small
+                 @click="setDownloadHref"
+                 :download="downloadFileName"
+                 :href="downloadJsonHref">
+            Download as json
+          </v-btn>
         </v-col>
       </v-row>
     </v-form>
@@ -80,14 +85,15 @@
   import i18n from '@/i18n'
   import ResizableContainer from '@/components/shared/ResizableContainer'
   import PrintPretty from '@/components/shared/PrintPretty'
-  import { HTTP_METHODS, REQUEST_DEFAULT_HEADERS } from '@/consts'
+  import { HTTP_METHODS, IDB_TABLE_NAMES, REQUEST_DEFAULT_HEADERS } from '@/consts'
   import { buildFetchAuthHeader } from '@/helpers'
   import { vuexAccessors } from '@/helpers/store'
   import { computed, ref } from '@vue/composition-api'
   import { showErrorSnackbar } from '@/mixins/ShowSnackbar'
   import { parseJsonBigInt } from '@/helpers/json_parse'
+  import { useIdb } from '@/services/IdbConnection'
   import RestQueryHistory from '@/components/Query/RestQueryHistory'
-  import { useDb } from '@/services/IdbConnection'
+  import RestQueryExamples from '@/components/Query/RestQueryExamples'
 
   export default {
     name: 'rest',
@@ -95,6 +101,7 @@
       PrintPretty,
       RestQueryHistory,
       ResizableContainer,
+      RestQueryExamples,
       'code-editor': () => ({
         component: import(/* webpackChunkName: "code-editor" */ '@/components/shared/CodeEditor')
       })
@@ -107,15 +114,8 @@
         vertical
       } = vuexAccessors('queryRest', ['method', 'path', 'requestBody', 'vertical'])
       const loading = ref(false)
-      const responseBody = ref({})
+      const responseBody = ref('')
       const responseStatus = ref(null)
-      const responseCaption = computed(() => {
-        if (responseStatus.value) {
-          return responseStatus.value
-        } else {
-          return i18n.t('rest.result')
-        }
-      })
 
       const canSendBody = computed(() => {
         return method.value !== 'GET' && method.value !== 'HEAD'
@@ -148,7 +148,7 @@
         return fetchOptions
       }
 
-      const { connection } = useDb('rest')
+      const { connection } = useIdb(IDB_TABLE_NAMES.REST)
       const setupDb = async () => await connection.initialize()
       setupDb()
 
@@ -163,11 +163,15 @@
           })
           .then(text => {
             loading.value = false
-            responseBody.value = parseJsonBigInt(text)
+            if (text) {
+              responseBody.value = parseJsonBigInt(text)
+            } else {
+              responseBody.value = ''
+            }
             if (responseStatus.value.toString().match(/^2\d\d/)) {
               connection.dbInsert({
                 method: method.value,
-                url: path.value,
+                path: path.value,
                 body: canSendBody.value ? requestBody.value : undefined,
                 favorite: 0,
                 date: new Date()
@@ -186,32 +190,8 @@
         exec: loadData
       }]
 
-      const loadCatExample = () => {
-        method.value = 'GET'
-        requestBody.value = '{\r\n\t"h": ["health", "index", "docs.count"]\r\n}'
-        path.value = '_cat/indices'
-        responseBody.value = ''
-        responseStatus.value = ''
-      }
-
-      const loadCreateExample = () => {
-        method.value = 'PUT'
-        requestBody.value = '{\r\n\t"settings": {\r\n\t\t"index": {\r\n\t\t\t"number_of_shards": 3,\r\n\t\t\t"number_of_replicas": 2\r\n\t\t}\r\n\t}\r\n}'
-        path.value = 'example_test_index'
-        responseBody.value = ''
-        responseStatus.value = ''
-      }
-
-      const loadDeleteExample = () => {
-        method.value = 'DELETE'
-        requestBody.value = '{}'
-        path.value = 'example_test_index'
-        responseBody.value = ''
-        responseStatus.value = ''
-      }
-
       const resetForm = () => {
-        method.value = 'GET'
+        method.value = HTTP_METHODS[1]
         requestBody.value = '{}'
         path.value = ''
         responseBody.value = ''
@@ -221,15 +201,40 @@
       const setRequest = item => {
         method.value = item.method
         requestBody.value = item.body
-        path.value = item.url
+        path.value = item.path
         responseBody.value = ''
         responseStatus.value = ''
+      }
+
+      const responseStatusClass = computed(() => {
+        if (!responseStatus.value || responseStatus.value.length === 0) return 'grey'
+
+        if (responseStatus.value.match(/^2/)) {
+          return 'green black--text'
+        } else if (responseStatus.value.match(/^3|4/)) {
+          return 'darken-2 yellow black--text'
+        } else if (responseStatus.value.match(/^5/)) {
+          return 'red black--text'
+        }
+
+        return 'grey'
+      })
+
+      const historyCollapsed = ref(false)
+
+      const downloadJsonHref = ref('#')
+      const downloadFileName = computed(() => {
+        return `${method.value.toLowerCase()}_${path.value.replace(/[\W_]+/g, '_')}.json`
+      })
+      const setDownloadHref = () => {
+        const value = typeof responseBody.value === 'string' ? responseBody.value : JSON.stringify(responseBody.value)
+        downloadJsonHref.value = `data:application/json,${encodeURIComponent(value)}`
       }
 
       return {
         loading,
         responseBody,
-        responseCaption,
+        responseStatus,
         vertical,
         path,
         method,
@@ -239,11 +244,13 @@
         canSendBody,
         formValid,
         loadData,
-        loadCatExample,
-        loadCreateExample,
-        loadDeleteExample,
         resetForm,
-        setRequest
+        setRequest,
+        responseStatusClass,
+        historyCollapsed,
+        downloadFileName,
+        downloadJsonHref,
+        setDownloadHref
       }
     }
   }