+++ src/mixins/History.js
@@ -1,36 +1,20 @@
-import { useDb } from '@/services/IdbConnection'
+import { useIdb } from '@/services/IdbConnection'
 import { ref, watch } from '@vue/composition-api'
 import { useAsyncFilter } from '@/mixins/UseAsyncTableFilter'
 import { debounce } from '@/helpers'
-
-const IDB_TABLE_NAMES = {
-  REST: 'rest',
-  SEARCH: 'search'
-}
-const IDB_TABLE_DEFINITIONS = {
-  [IDB_TABLE_NAMES.REST]: {
-    indexes: ['date'],
-    filterableColumns: ['method', 'url']
-  },
-  [IDB_TABLE_NAMES.SEARCH]: {
-    indexes: ['date'],
-    filterableColumns: ['url']
-  }
-}
+import { IDB_TABLE_DEFINITIONS } from '@/consts'
 
 export const useHistory = tableName => {
-  const { connection } = useDb(tableName)
+  const { connection } = useIdb(tableName)
   connection.initialize()
-  const selectedItem = ref(null)
 
-  const setSelectedItem = item => (selectedItem.value = item)
   const favoriteItem = item => connection.dbUpdate(Object.assign({}, item, { favorite: !item.favorite ? 1 : 0 }))
   const removeItem = id => connection.dbDelete(id)
-  const clear = () => {
+  const clearAll = () => {
     if (confirm('Are you sure? This will remove all entries from your history.')) connection.dbClear()
   }
   const clearNonFavorites = () => {
-    connection.dbClearNonFavorites()
+    return connection.dbClearNonFavorites()
   }
 
   const filter = ref('')
@@ -44,18 +28,16 @@ export const useHistory = tableName => {
     if (onlyFavorites.value) results = results.filter(entry => entry.favorite === 1)
     items.value = await asyncFilterTable(results, filter.value, IDB_TABLE_DEFINITIONS[tableName].filterableColumns)
   }
-  const debouncedFilterTable = debounce(filterTable, 350)
+  const debouncedFilterTable = debounce(filterTable, 250)
   watch(filter, debouncedFilterTable)
   watch([onlyFavorites, connection.entries], filterTable)
 
   return {
-    selectedItem,
     connection,
     loading: connection.loading,
-    setSelectedItem,
     favoriteItem,
     removeItem,
-    clear,
+    clearAll,
     clearNonFavorites,
     items,
     filter,